TTC: Changed region rules + added thwomp rule
For some reason my Upper/Top rules were outdated so I updated those and adjusted as needed.

I was able to get to the Top of TTC with any combination of CL and LG/TJ/BF/WK doing the intended route with time slowed (which is forced if randomized to a secret and easily selectable if it can be adjusted) and time stop. 

Getting to the Thwomp in the non-strict scenarios may require a different moveset than getting to the top the clock to be moving to ride the hand, but as this is likely the intended way to get the star I do not think this is an issue. If riding the hand (non-strict route), ledge grab and wall kick do not work for getting on the thwomp so I added a separate rule for the star to exclude those.